DB설정 추가

+++ src/main/java/com/example/member/domain/Member.java
@@ -1,16 +1,15 @@
 package com.example.member.domain;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
 import lombok.Data;
-import org.springframework.data.annotation.Id;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
 
 @Entity
 @Data
 public class Member {
 
     @Id
-    private Long id;
+    private Integer id;
 
     private String name;
 }
